재구성: tag 컴포넌트 분리. 코드 간결화.
